refactor: ファイルに関するユーティリティを別モジュールへ分離
